Keep client test browser shims easier to extend
Extract the browser-mock builders from the shared client test helpers so
DOM environment setup stays focused on composition while the mock
navigator/history/location logic remains isolated and reusable.

Constraint: Preserve client integration helper behavior used by prefetch and router tests
Constraint: Repo pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Splitting the full DOMEnvironment harness into multiple modules | broader churn than this queue item needs
Confidence: high
Scope-risk: narrow
Directive: Keep browser-mocks.ts limited to pure browser shim builders; leave DOM environment orchestration in test-helpers.ts
Tested: deno test --no-check --allow-all tests/integration/client/prefetch.test.ts tests/integration/client/router.test.ts; deno fmt --check tests/integration/client/test-helpers.ts tests/integration/client/browser-mocks.ts; deno lint tests/integration/client/test-helpers.ts tests/integration/client/browser-mocks.ts; git diff --check
Not-tested: Pre-commit verify:quick (fails on existing CLI boundary violations in cli/commands/extension/*.ts)
